started converting file to ts
